+++ com.unity.toonshader/CHANGELOG.md
@@ -50,6 +50,8 @@
 * Renamed `MatCap Power on Shadow` to `MatCap Blending on Shadows`.
 * Renamed `Color Shift Speed` to `Color Shifting Speed`.
 * Renamed `Blur Level of ShadingGradeMap` to `ShadingGradeMap Blur Level`.
+* Renamed `MatCap Sampler` to `MatCap  Map`.
+* Renamed `Blur Level of MatCap Sampler` to `MatCap Blur Level`.
 ### Added:
 * Legacy: Compatibility with Single Pass Stereo Rendering.
 * Legacy: Graphics Test 2022.2.